Refactor openapi public and trial into platform
