--- openapi/api_public.template.yaml
@@ -1,130 +1,0 @@
-swagger: "2.0"
-info:
-  version: 16.7.0
-  title: Public API
-  description: Mobile and web proxy API gateway.
-host: localhost
-basePath: /
-schemes:
-  - https
-paths:
-  "/":
-      x-swagger-router-controller: ServerInfoController
-      get:
-        operationId: Redirect
-        summary: Redirect
-        description: |
-          Redirects to a specific URL.
-  
-        responses:
-          "302":
-            description: HTTP Redirect.
-            headers:
-              Location:
-                description: URL where the client is being redirected.
-                type: string
-  "/info":
-    x-swagger-router-controller: ServerInfoController
-    get:
-      operationId: getServerInfo
-      summary: Runtime server info
-      description: >
-        Returns runtime information about the server.
-
-      responses:
-        "200":
-          description: Runtime server info.
-          schema:
-            $ref: "#/definitions/ServerInfo"
-          examples:
-            application/json:
-              version: "0.0.1"
-              min_app_version: "0.0.0"
-              min_app_version_pagopa: "0.0.0"
-  "/api/v1/ping":
-    x-swagger-router-controller: StatusServicesController
-    get:
-      operationId: getPing
-      summary: Check online status
-      description: >
-        Returns no content.
-      responses:
-        "204":
-          description: Service is alive.
-  "/api/v1/status":
-    x-swagger-router-controller: StatusServicesController
-    get:
-      summary: Get service statuses
-      description: Returns the current status of all microservices.
-      responses:
-        '200':
-          description: App backend microservices statuses.
-          schema:
-            $ref: '#/definitions/ServicesStatus'
-        '500':
-          description: Internal Server Error
-          schema:
-            $ref: "#/definitions/ProblemJson"
-definitions:
-  ProblemJson:
-    $ref: "../node_modules/@pagopa/io-functions-commons/openapi/definitions.yaml#/ProblemJson"
-  ServerInfo:
-    type: object
-    title: Server information
-    properties:
-      version:
-        type: string
-      min_app_version:
-        $ref: "#/definitions/VersionPerPlatform"
-      min_app_version_pagopa:
-        $ref: "#/definitions/VersionPerPlatform"
-    required:
-      - version
-      - min_app_version
-      - min_app_version_pagopa
-  VersionPerPlatform:
-    type: object
-    title: Specify a version for ios and another for android
-    properties:
-      ios:
-        type: string
-      android:
-        type: string
-    required:
-      - ios
-      - android
-  ServicesStatus:
-    type: object
-    properties:
-      services:
-        type: array
-        items:
-          $ref: "#/definitions/ServiceStatus"
-    required:
-      - services
-  ServiceStatus:
-    type: object
-    properties:
-      id:
-        type: string
-      status:
-        $ref: '#/definitions/StatusEnum'
-      details:
-        type: string
-    required:
-     - id
-     - status
-     - details
-  StatusEnum:
-    type: string
-    description: Status of a service
-    enum:
-      - OK
-      - WARNING
-      - CRITICAL
-      - PARTIAL_AVAILABILITY
-      - MANTAINANCE
-consumes:
-  - application/json
-produces:
-  - application/json
