--- openapi/api_trial_system.template.yaml
@@ -1,125 +1,0 @@
-swagger: "2.0"
-info:
-  version: 16.7.0
-  title: Trial System API
-host: localhost
-basePath: /api/v1
-schemes:
-  - https
-security:
-  - Bearer: []
-securityDefinitions:
-  Bearer:
-    type: apiKey
-    name: Authorization
-    in: header
-paths:
-  "/trials/{trialId}/subscriptions":
-    parameters:
-      - $ref: "#/parameters/pathTrialId"
-    get:
-      operationId: getSubscription
-      summary: Get subscription detail
-      description: >-
-        Returns the details about the subscription of the authenticated user for
-        the specified trial.
-      responses:
-        "200":
-          description: Success.
-          schema:
-            $ref: "#/definitions/Subscription"
-        "401":
-          description: Authorization error.
-        "404":
-          description: |
-            Not Found. This occurs when the specified user is not listed under the trial.
-        "429":
-          description: Too may requests.
-        "500":
-          description: There was an error.
-          schema:
-            $ref: "#/definitions/ProblemJson"
-    post:
-      operationId: createSubscription
-      summary: Create a subscription
-      description: >-
-        Subscribe the authenticated user to the given trial.
-      responses:
-        "201":
-          description: |
-            Created: the request has been processed and the user has been registered for the trial.
-          schema:
-            $ref: "#/definitions/Subscription"
-        "202":
-          description: |
-            Accepted: the request has been received, but the user may not yet be subscribed to the trial.
-            The server will process the request asynchronously.
-            Resubmitting the same request may result in a `409 Conflict` error.
-        "400":
-          description: Validation error.
-          schema:
-            $ref: "#/definitions/ProblemJson"
-        "401":
-          description: Authorization error.
-        "409":
-          description: The resource already exists.
-        "429":
-          description: Too may requests.
-        "500":
-          description: There was an error.
-          schema:
-            $ref: "#/definitions/ProblemJson"
-
-parameters:
-  pathTrialId:
-    name: trialId
-    in: path
-    required: true
-    type: string
-    minLength: 1
-
-definitions:
-  ProblemJson:
-    $ref: "../node_modules/@pagopa/io-functions-commons/openapi/definitions.yaml#/ProblemJson"
-  CreatedAt:
-    $ref: "../node_modules/@pagopa/io-functions-commons/openapi/definitions.yaml#/Timestamp"
-  SubscriptionState:
-    type: string
-    description: |-
-      - UNSUBSCRIBED: The user is not subscribed to the specified trial and does
-          not have any access to the specified trial.
-      - SUBSCRIBED: The user is subscribed to the specified trial and does
-          not have any access to the specified trial. The system is going to
-          activate the user if any slot for the trial is available.
-      - ACTIVE: The user has the access to the specified trial.
-      - DISABLED: The user does not have the access to the specified trial.
-
-      When a user subscribes a trial, the subscription is in the `SUBSCRIBED` state
-      If the user wants to unsubscribe, the state changes to `UNSUBSCRIBED`.
-      When the trial is going to be activated (by the Trial Manager, the owner of the trial), then the state of the
-      selected `SUBSCRIBED` users becomes `ACTIVE`.
-      The Trial Manager can change the state of any user, within its own subscription,
-      to any state (e.g. can disable a subscription setting the state to `DISABLED`).
-    enum:
-      - UNSUBSCRIBED
-      - SUBSCRIBED
-      - ACTIVE
-      - DISABLED
-  TrialId:
-    type: string
-    description: >-
-      Unique identifier of the trial.
-    minLength: 1
-  Subscription:
-    type: object
-    required:
-      - trialId
-      - state
-      - createdAt
-    properties:
-      trialId:
-        $ref: '#/definitions/TrialId'
-      state:
-        $ref: '#/definitions/SubscriptionState'
-      createdAt:
-        $ref: '#/definitions/CreatedAt'
