+++ package.json
@@ -34,10 +34,9 @@
     "generate:proxy:notification-models": "rimraf generated/notifications && gen-api-models --api-spec ./openapi/generated/api_notifications.yaml --out-dir generated/notifications",
     "generate:messages:notifications": "rimraf generated/messages && gen-api-models --api-spec notification_queue_messages.yaml --out-dir generated/messages",
     "generate:proxy:myportal-models": "rimraf generated/myportal && gen-api-models --api-spec api_myportal.yaml --out-dir generated/myportal",
-    "generate:proxy:trial-system-models": "rimraf generated/trial-system && gen-api-models --api-spec api_trial_system.yaml --out-dir generated/trial-system",
     "generate:api:trial-system": "rimraf generated/trial-system-api && gen-api-models --api-spec https://raw.githubusercontent.com/pagopa/trial-system/081d98c37a116c9d842f31c54a2fe641680c5424/infra/resources/modules/commons/api/ts_management/v1/_openapi.yaml --no-strict --out-dir generated/trial-system-api --request-types --response-decoders --client",
     "generate:proxy:first-lc-models": "rimraf generated/first-lc-proxy-models && gen-api-models --api-spec api_lollipop_first_consumer.yaml --out-dir generated/first-lc-proxy-models",
-    "generate:proxy:public-models": "rimraf generated/public  && gen-api-models --api-spec api_public.yaml --out-dir generated/public",
+    "generate:proxy:platform-models": "rimraf generated/platform && gen-api-models --api-spec ./openapi/generated/api_platform.yaml --out-dir generated/platform",
     "generate:api:io-messages": "rimraf generated/io-messages-api && gen-api-models --api-spec https://raw.githubusercontent.com/pagopa/io-functions-app-messages/v5.3.0/openapi/index.yaml --no-strict --out-dir generated/io-messages-api --request-types --response-decoders --client",
     "generate:proxy:cgn-models": "rimraf generated/cgn && gen-api-models --api-spec ./openapi/generated/api_cgn.yaml --out-dir generated/cgn",
     "generate:proxy:io-sign-models": "rimraf generated/io-sign && gen-api-models --api-spec api_io_sign.yaml --out-dir generated/io-sign",
