+++ src/__tests__/api-specs.test.ts
@@ -10,7 +10,7 @@ describe("API proxy specs", () => {
 });
 
 describe("API notifications specs", () => {
-  const specFilePath = `${__dirname}/../../api_notifications.yaml`;
+  const specFilePath = `${__dirname}/../../openapi/generated/api_notifications.yaml`;
 
   it("should be valid", async () => {
     const api = await SwaggerParser.bundle(specFilePath);
